requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.5
python-dateutil>=2.8.2
h3>=3.7.6
numpy>=1.24.0  # Required by pandas and h3 